requirements.txt
Flask
flask-cors
flask-bcrypt
mysql-connector-python
gunicorn
fpdf
psycopg2-binary